refactor: UserController and AuthController
